# packages/schemas/index.ts
/**
 * @go2asia/schemas
 * 
 * Zod схемы для валидации данных
 */

import { z } from 'zod';

// Единый формат ошибки API
export const ErrorResponseSchema = z.object({
  error: z.object({
    code: z.string(),
    message: z.string(),
    traceId: z.string(),
    key: z.string().optional(),
  }),
});

export type ErrorResponse = z.infer<typeof ErrorResponseSchema>;

// Реэкспорт для использования в других пакетах
export * from 'zod';
